requirements.txt
google-cloud-documentai==2.25.0
requests==2.28.1